+++ src/Container.php
@@ -568,13 +568,18 @@ class Container implements ArrayAccess, ContainerInterface
      * @param  string[]|null                  $afterBuildMethods  An array of methods that should be called on the
      *                                                            instance after it has been built; the methods should
      *                                                            not require any argument.
+     * @param bool $afterBuildAll                                 Whether to call the after build methods on only the
+     *                                                            base instance or all instances of the decorator chain.
      *
      * @return void This method does not return any value.
      * @throws ContainerException
      */
-    public function singletonDecorators($id, $decorators, array $afterBuildMethods = null)
+    public function singletonDecorators($id, $decorators, array $afterBuildMethods = null, $afterBuildAll = false)
     {
-        $this->resolver->singleton($id, $this->getDecoratorBuilder($decorators, $id, $afterBuildMethods));
+        $this->resolver->singleton(
+            $id,
+            $this->getDecoratorBuilder($decorators, $id, $afterBuildMethods, $afterBuildAll)
+        );
     }
 
     /**
@@ -584,12 +589,20 @@ class Container implements ArrayAccess, ContainerInterface
      * @param string                        $id                The id to bind the decorator tail to.
      * @param array<string>|null            $afterBuildMethods A set of method to run on the built decorated instance
      *                                                         after it's built.
+     * @param bool $afterBuildAll                              Whether to run the after build methods only on the base
+     *                                                         instance (default, false) or on all instances of the
+     *                                                         decorator chain.
+     *
      * @return BuilderInterface The callable or Closure that will start building the decorator chain.
      *
      * @throws ContainerException If there's any issue while trying to register any decorator step.
      */
-    private function getDecoratorBuilder(array $decorators, $id, array $afterBuildMethods = null)
-    {
+    private function getDecoratorBuilder(
+        array $decorators,
+        $id,
+        array $afterBuildMethods = null,
+        $afterBuildAll = false
+    ) {
         $decorator = array_pop($decorators);
 
         if ($decorator === null) {
@@ -600,7 +613,9 @@ class Container implements ArrayAccess, ContainerInterface
             $previous = isset($builder) ? $builder : null;
             $builder = $this->builders->getBuilder($id, $decorator, $afterBuildMethods, $previous);
             $decorator = array_pop($decorators);
-            $afterBuildMethods = [];
+            if (!$afterBuildAll) {
+                $afterBuildMethods = [];
+            }
         } while ($decorator !== null);
 
         return $builder;
@@ -616,15 +631,17 @@ class Container implements ArrayAccess, ContainerInterface
      *                                                            should be bound to.
      * @param  array<string|object|callable>  $decorators         An array of implementations that decorate an object.
      * @param  string[]|null                  $afterBuildMethods  An array of methods that should be called on the
-     *                                                            instance after it has been built; the methods should
-     *                                                            not require any argument.
+     *                                                            base instance after it has been built; the methods
+     *                                                            should not require any argument.
+     * @param bool $afterBuildAll                                 Whether to call the after build methods on only the
+     *                                                            base instance or all instances of the decorator chain.
      *
      * @return void This method does not return any value.
      * @throws ContainerException If there's any issue binding the decorators.
      */
-    public function bindDecorators($id, array $decorators, array $afterBuildMethods = null)
+    public function bindDecorators($id, array $decorators, array $afterBuildMethods = null, $afterBuildAll = false)
     {
-        $this->resolver->bind($id, $this->getDecoratorBuilder($decorators, $id, $afterBuildMethods));
+        $this->resolver->bind($id, $this->getDecoratorBuilder($decorators, $id, $afterBuildMethods, $afterBuildAll));
     }
 
     /**